requirements.txt
streamlit 
tensorflow 
pillow 
numpy
pytesseract 
